Replace invalid quota in SQLSelectItem, alias and SPLPropertyExpr and modify starrocks output visitor.
