+++ core/src/main/java/com/alibaba/druid/sql/dialect/starrocks/ast/statement/StarRocksCreateTableStatement.java
@@ -2,8 +2,6 @@ package com.alibaba.druid.sql.dialect.starrocks.ast.statement;
 
 import com.alibaba.druid.DbType;
 import com.alibaba.druid.sql.ast.DistributedByType;
-import com.alibaba.druid.sql.ast.SQLExpr;
-import com.alibaba.druid.sql.ast.SQLOrderBy;
 import com.alibaba.druid.sql.ast.statement.SQLAssignItem;
 import com.alibaba.druid.sql.ast.statement.SQLCreateTableStatement;
 import com.alibaba.druid.sql.ast.statement.SQLSelectOrderByItem;
@@ -15,24 +13,11 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class StarRocksCreateTableStatement extends SQLCreateTableStatement implements StarRocksObject {
-    protected SQLExpr engine;
-    protected SQLOrderBy orderBy;
     protected DistributedByType distributedByType;
     protected final List<SQLSelectOrderByItem> distributedBy = new ArrayList<>();
     protected final List<SQLAssignItem> brokerProperties = new ArrayList<>();
     public StarRocksCreateTableStatement() {
         super(DbType.starrocks);
-    }
-
-    public SQLExpr getEngine() {
-        return engine;
-    }
-
-    public void setEngine(SQLExpr x) {
-        if (x != null) {
-            x.setParent(this);
-        }
-        this.engine = x;
     }
 
     public DistributedByType getDistributedByType() {
@@ -49,17 +34,6 @@ public class StarRocksCreateTableStatement extends SQLCreateTableStatement imple
 
     public List<SQLAssignItem> getBrokerProperties() {
         return brokerProperties;
-    }
-
-    public SQLOrderBy getOrderBy() {
-        return orderBy;
-    }
-
-    public void setOrderBy(SQLOrderBy orderBy) {
-        if (orderBy != null) {
-            orderBy.setParent(this);
-        }
-        this.orderBy = orderBy;
     }
 
     @Override
