+++ core/src/main/resources/META-INF/druid/parser/clickhouse/dialect.properties
@@ -1,0 +1,1 @@
+quote=`
