+++ core/src/main/resources/META-INF/druid/parser/h2/dialect.properties
@@ -1,0 +1,1 @@
+quote=`
