+++ core/src/test/resources/bvt/parser/starrocks/0.txt
@@ -1,3 +1,19 @@
+select "a"."b"
+--------------------
+SELECT `a`.`b`
+------------------------------------------------------------------------------------------------------------------------
+select "a" as "b"
+--------------------
+SELECT `a` AS `b`
+------------------------------------------------------------------------------------------------------------------------
+create table test(
+    "test" int
+)
+--------------------
+CREATE TABLE test (
+	`test` int
+)
+------------------------------------------------------------------------------------------------------------------------
 select a from "test".test
 --------------------
 SELECT a
